+++ apps/web/next.config.mjs
@@ -55,7 +55,6 @@ const config = {
     '@pancakeswap/ui',
     '@pancakeswap/uikit',
     '@pancakeswap/farms',
-    '@pancakeswap/pools',
     '@pancakeswap/localization',
     '@pancakeswap/hooks',
     '@pancakeswap/utils',
@@ -185,7 +184,7 @@ const config = {
         source: '/images/tokens/:address',
         destination: 'https://tokens.pancakeswap.finance/images/:address',
         permanent: false,
-      }
+      },
     ]
   },
   webpack: (webpackConfig, { webpack, isServer }) => {