+++ packages/uikit/src/widgets/Menu/components/footerConfig.ts
@@ -34,6 +34,14 @@ export const footerLinks: (t: ContextApi["t"]) => FooterLinkType[] = (t) => [
         label: t("Terms Of Service"),
         href: "https://pancakeswap.finance/terms-of-service",
       },
+      {
+        label: "—",
+      },
+      {
+        label: t("Merchandise"),
+        href: "https://merch.pancakeswap.finance/",
+        isHighlighted: true,
+      },
     ],
   },
   {