+++ apis/routing/CHANGELOG.md
@@ -1,5 +1,56 @@
 # routing-api
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [24f51d314]
+  - @pancakeswap/smart-router@4.6.0
+
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [e9c080787]
+  - @pancakeswap/smart-router@4.5.2
+
+## 0.1.2
+
+### Patch Changes
+
+- @pancakeswap/farms@1.0.10
+- @pancakeswap/smart-router@4.5.1
+- @pancakeswap/v3-sdk@3.3.1
+
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [659be0529]
+  - @pancakeswap/smart-router@4.5.0
+
+## 0.1.0
+
+### Minor Changes
+
+- 5e15c611e: Add linea support
+
+### Patch Changes
+
+- Updated dependencies [5e15c611e]
+  - @pancakeswap/smart-router@4.4.0
+  - @pancakeswap/sdk@5.3.0
+  - @pancakeswap/v3-sdk@3.3.0
+  - @pancakeswap/farms@1.0.9
+
+## 0.0.13
+
+### Patch Changes
+
+- @pancakeswap/farms@1.0.8
+- @pancakeswap/smart-router@4.3.6
+- @pancakeswap/v3-sdk@3.2.3
+
 ## 0.0.12
 
 ### Patch Changes