+++ packages/awgmi/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @pancakeswap/awgmi
 
+## 0.1.11
+
+### Patch Changes
+
+- @pancakeswap/utils@4.0.1
+
+## 0.1.10
+
+### Patch Changes
+
+- @pancakeswap/utils@4.0.0
+
+## 0.1.9
+
+### Patch Changes
+
+- 299cf46b7: Upgrade blocto sdk version
+  - @pancakeswap/utils@3.0.3
+
 ## 0.1.8
 
 ### Patch Changes