+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @pancakeswap/hooks
 
+## 0.0.12
+
+### Patch Changes
+
+- @pancakeswap/utils@4.0.1
+
+## 0.0.11
+
+### Patch Changes
+
+- @pancakeswap/utils@4.0.0
+
+## 0.0.10
+
+### Patch Changes
+
+- @pancakeswap/utils@3.0.3
+
 ## 0.0.9
 
 ### Patch Changes