+++ packages/localization/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @pancakeswap/localization
 
+## 4.0.1
+
+### Patch Changes
+
+- @pancakeswap/utils@4.0.1
+- @pancakeswap/hooks@0.0.12
+
+## 4.0.0
+
+### Patch Changes
+
+- @pancakeswap/utils@4.0.0
+- @pancakeswap/hooks@0.0.11
+
+## 3.0.3
+
+### Patch Changes
+
+- @pancakeswap/utils@3.0.3
+- @pancakeswap/hooks@0.0.10
+
 ## 3.0.2
 
 ### Patch Changes