+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @pancakeswap/utils
 
+## 4.0.1
+
+### Patch Changes
+
+- Updated dependencies [d0f9b28a9]
+  - @pancakeswap/tokens@0.3.1
+  - @pancakeswap/awgmi@0.1.11
+  - @pancakeswap/localization@4.0.1
+  - @pancakeswap/token-lists@0.0.8
+
+## 4.0.0
+
+### Patch Changes
+
+- Updated dependencies [5e15c611e]
+  - @pancakeswap/tokens@0.3.0
+  - @pancakeswap/awgmi@0.1.10
+  - @pancakeswap/localization@4.0.0
+  - @pancakeswap/token-lists@0.0.8
+
+## 3.0.3
+
+### Patch Changes
+
+- Updated dependencies [299cf46b7]
+  - @pancakeswap/awgmi@0.1.9
+  - @pancakeswap/localization@3.0.3
+  - @pancakeswap/token-lists@0.0.8
+  - @pancakeswap/tokens@0.2.3
+
 ## 3.0.2
 
 ### Patch Changes