+++ packages/v3-sdk/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @pancakeswap/v3-sdk
 
+## 3.3.1
+
+### Patch Changes
+
+- Updated dependencies [d0f9b28a9]
+  - @pancakeswap/tokens@0.3.1
+
+## 3.3.0
+
+### Minor Changes
+
+- 5e15c611e: Add linea support
+
+### Patch Changes
+
+- Updated dependencies [5e15c611e]
+  - @pancakeswap/sdk@5.3.0
+  - @pancakeswap/tokens@0.3.0
+
+## 3.2.3
+
+### Patch Changes
+
+- @pancakeswap/tokens@0.2.3
+
 ## 3.2.2
 
 ### Patch Changes